requirements.txt
# LocalRAG_Pro - 依存パッケージ
# インストール: pip install -r requirements.txt

# ===== Web API =====
flask>=3.0.0
flask-cors>=4.0.0

# ===== RAG エンジン =====
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.3.0
llama-index-embeddings-huggingface>=0.3.0

# ===== Embedding モデル =====
sentence-transformers>=3.0.0
torch>=2.0.0

# ===== ドキュメント読み込み =====
pymupdf>=1.24.0          # PDF テキスト抽出
python-docx>=1.1.0       # DOCX
docx2txt>=0.8            # DOCX フォールバック
python-pptx>=1.0.0       # PPTX
openpyxl>=3.1.0          # XLSX

# ===== OCR（オプション） =====
# Tesseract 本体は別途インストールが必要
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
pdf2image>=1.17.0        # PDF → 画像変換（poppler 必要）
pytesseract>=0.3.13      # OCR

# ===== テスト =====
pytest>=8.0.0
pytest-cov>=5.0.0
